refactor: 프로젝트 구조화
