+++ src/js/racingcar.js
@@ -1,0 +1,8 @@
+import RacingCarController from "./racingcar/controller.js";
+class RacingCar {
+  constructor() {
+    new RacingCarController();
+  }
+}
+
+export default RacingCar;
